Refactor likelihood module: In 'maximization_inputs.py', import from two modules with likelihood functions.
